fix incorrect timing
function intrinsic `cpu_time` is not thread aware and therefore if the
application is threaded (which some of the benchmarks are) then it can
give unpredictable timings.

OMP's `omp_get_wtime` should be used instead. This can correctly measure
the waltime when threading is used.

I have also added a new script run_benchmarks.sh that will run the
benchmarks and save their output to file.